+++ Classes/ImageService/GumletImageService.php
@@ -1,0 +1,88 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SomehowDigital\Typo3\MediaProcessing\ImageService;
+
+use SomehowDigital\Typo3\MediaProcessing\UriBuilder\GumletUri;
+use SomehowDigital\Typo3\MediaProcessing\UriBuilder\UriSourceInterface;
+use TYPO3\CMS\Core\Imaging\ImageDimension;
+use TYPO3\CMS\Core\Resource\Processing\TaskInterface;
+
+class GumletImageService extends ImageServiceAbstract
+{
+	public static function getIdentifier(): string
+	{
+		return 'gumlet';
+	}
+
+	public function __construct(
+		protected readonly string $endpoint,
+		protected readonly UriSourceInterface $source,
+	) {
+	}
+
+	public function getEndpoint(): string
+	{
+		return $this->endpoint;
+	}
+
+	public function hasConfiguration(): bool
+	{
+		return filter_var($this->getEndpoint(), FILTER_VALIDATE_URL) !== false;
+	}
+
+	public function canProcessTask(TaskInterface $task): bool
+	{
+		return
+			$task->getSourceFile()->exists() &&
+			$task->getSourceFile()->getStorage()?->isPublic() &&
+			in_array($task->getName(), ['Preview', 'CropScaleMask'], true) &&
+			in_array($task->getSourceFile()->getMimeType(), [
+				'image/jpeg',
+				'image/jpx',
+				'image/jpm',
+				'image/jxl',
+				'image/png',
+				'image/webp',
+				'image/tiff',
+				'application/pdf',
+				'image/gif',
+				'image/heic',
+				'image/heif',
+				'image/avif',
+			]);
+	}
+
+	public function processTask(TaskInterface $task): ImageServiceResult
+	{
+		$file = $task->getSourceFile();
+		$configuration = $task->getTargetFile()->getProcessingConfiguration();
+		$dimension = ImageDimension::fromProcessingTask($task);
+
+		$uri = new GumletUri($this->getEndpoint());
+		$uri->setSource($this->source->getSource($file));
+
+		if (isset($configuration['crop'])) {
+			$uri->setCrop(
+				(int) $configuration['crop']->getWidth(),
+				(int) $configuration['crop']->getHeight(),
+				(int) $configuration['crop']->getOffsetLeft(),
+				(int) $configuration['crop']->getOffsetTop(),
+			);
+		}
+
+		if (isset($configuration['width']) || isset($configuration['maxWidth'])) {
+			$uri->setWidth((int) ($configuration['width'] ?? $configuration['maxWidth']));
+		}
+
+		if (isset($configuration['height']) || isset($configuration['maxHeight'])) {
+			$uri->setHeight((int) ($configuration['height'] ?? $configuration['maxHeight']));
+		}
+
+		return new ImageServiceResult(
+			$uri,
+			$dimension,
+		);
+	}
+}